% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSpecific.R
\name{bioGender}
\alias{bioGender}
\title{Detects gender of microarray samples}
\usage{
bioGender(x, geneColName = "Gene.Symbol", probeColName = "Probe",
  maleGenes = c("RPS4Y1", "KDM5D"), femaleGenes = "XIST")
}
\arguments{
\item{x}{Expression data frame with gene symbol and probename information in the first columns}

\item{geneColName}{column name for gene symbols}

\item{probeColName}{column name for probeset names}

\item{maleGenes}{names of male genes}

\item{femaleGenes}{names of female genes}
}
\description{
Detects gender of microarray samples based on expression of sex specific genes as described by Toker et al. 2016 (PMC5034794).
}
