% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{cantor_pairing}
\alias{cantor_pairing}
\title{Cantor Pairing}
\usage{
cantor_pairing(x, y)
}
\arguments{
\item{x}{First integer}

\item{y}{Second integer}
}
\value{
Integer corresponding to the paired value
}
\description{
Performs cantor pairing on two integers
}
