% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R
\name{celFiles}
\alias{celFiles}
\title{List cel files}
\usage{
celFiles(..., listGzipped = FALSE)
}
\arguments{
\item{...}{arguments to pass to \code{list.files}}

\item{listGzipped}{logical. If true adds gzipped files to the results}
}
\value{
A string vector, listing files
}
\description{
list.celfiles from oligo package. it's better than affy since it 
has listGzipped option.
}

