% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R
\name{checkLines}
\alias{checkLines}
\title{Read specific lines from a file}
\usage{
checkLines(file, lines, fun = readLines, ...)
}
\arguments{
\item{file}{the name of the file which lines will be read from}

\item{fun}{function that will process each line}
}
\value{
A list containing the outputs per line
}
\description{
Creates a text connection for specific lines of a file and
returns them as an output of \code{fun}. Modified from  http://stackoverflow.com/questions/18235269/efficiently-reading-specific-lines-from-large-files-into-r
}

