% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R, R/ogbox2.R
\name{clc}
\alias{clc}
\title{Clear display}
\usage{
clc()

clc()
}
\description{
clear display, taken from http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r

clear display, taken from http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r
}

