% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{downloadGSE}
\alias{downloadGSE}
\title{Downloads an entire dataset from GEO}
\usage{
downloadGSE(GSE, celDir, metaDir)
}
\arguments{
\item{GSE}{GSE id}

\item{celDir}{directory to save the cel files}

\item{metaDir}{directory to save the soft file}
}
\description{
Downloads an entire dataset from GEO
}
