% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit.R
\name{explicit}
\alias{explicit}
\title{Make function calls explicit}
\usage{
explicit(fileIn, fileOut = fileIn, ignore = "base")
}
\arguments{
\item{fileIn}{input file}

\item{fileOut}{output file. default is to overwrite fileIn}

\item{ignore}{which packages to ignore when translating}
}
\description{
This function turns function calls into explicit function calls (packageName::functionName) in a file
The packages must be currently loaded
}

