% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{forkCRAN}
\alias{forkCRAN}
\title{Make a github repository from a specific version of a pacakge}
\usage{
forkCRAN(
  pkg,
  version = NULL,
  newname = NULL,
  token = NULL,
  private = FALSE,
  has_issues = FALSE,
  has_wiki = FALSE
)
}
\arguments{
\item{pkg}{Name of the package}

\item{version}{Desired version}

\item{newname}{Name of the resulting package. If NULL it'll default to packagename.versionnumber}

\item{token}{Github token if NULL GITHUB_PAT environment variable will be used}

\item{private, has_issues, has_wiki}{arguments to pass to gh}
}
\description{
Make a github repository from a specific version of a pacakge
}
