% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{forkOldCRAN}
\alias{forkOldCRAN}
\title{Make a github repository from a specific version of a pacakge}
\usage{
forkOldCRAN(pkg, version, token = NULL, private = FALSE,
  has_issues = FALSE, has_wiki = FALSE)
}
\arguments{
\item{pkg}{Name of the package}

\item{version}{Desired version}

\item{private, has_issues, has_wiki}{arguments to pass to gh}

\item{credentials}{}
}
\description{
Make a github repository from a specific version of a pacakge
}
