% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listManipulation.R
\name{frame2tree}
\alias{frame2tree}
\title{Transform a data frame into a nested list.}
\usage{
frame2tree(frame, levels = colnames(frame))
}
\value{
A nested list
}
\description{
Takes in given columns of a data frame and transforms them into a nested list based on the column names 
given. Eg. if first level is \code{c(1,1,2)} and second level is \code{c('a','b','c')}, it will return a list of 
length 2 that includes lists named 1 and 2. List 1 will include 2 empty lists named a and b while list 2 will include
1 empty list called c.
}

