% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{generic_install}
\alias{generic_install}
\title{Install generic remote package}
\usage{
generic_install(
  x,
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  repos = getOption("repos"),
  type = getOption("pkgType"),
  ...
)
}
\description{
Better version of the original by jimhester
https://github.com/r-lib/remotes/issues/383
}
