% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomNames.R
\name{getRandomNames}
\alias{getRandomNames}
\title{#' @export
get_random_name = function(retry){
    left = left()
    right = right()
  name = paste0(left[sample(x = length(left), size = 1)],'_',right[sample(x = length(right), size = 1)])
  if(name == 'boring_wozniak'){
    get_random_name(retry)
  }
  if(retry>0){
    name = paste0(name,sample(10,1))
  }
  return(name)
}}
\usage{
getRandomNames(n = 1, alliterate = FALSE)
}
\description{
#' @export
get_random_name = function(retry){
    left = left()
    right = right()
  name = paste0(left[sample(x = length(left), size = 1)],'_',right[sample(x = length(right), size = 1)])
  if(name == 'boring_wozniak'){
    get_random_name(retry)
  }
  if(retry>0){
    name = paste0(name,sample(10,1))
  }
  return(name)
}
}
