% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{getVersion}
\alias{getVersion}
\title{Get version of a local package}
\usage{
getVersion(pkg = ".")
}
\arguments{
\item{pkg}{Pakcage to check version of. Defaults to local directory}
}
\description{
Get version of a local package
}
