% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEO.R
\name{gseDown}
\alias{gseDown}
\title{Download multiple cel files from GSE}
\usage{
gseDown(GSE, regex = NULL, outDir, extension = ".cel", overwrite = F,
  unzip = T)
}
\arguments{
\item{GSE}{GSE identifier of the dataset}

\item{regex}{regular expression to limit files to be downloaded. Searches the regex inside GSM titles}

\item{outDir}{directory to save the cel files}

\item{extension}{which extension to use when saving the file}

\item{overwrite}{should it overwrite the files if they exist}

\item{unzip}{should it decompress downloaded files.}
}
\description{
Downloads all CEL files from GSMs related to a GSE. Optionally, a regular expression can be specified to download a subset of cel files
}

