% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latexImg}
\alias{latexImg}
\title{Creates a markdown link to the rendered form of a latex formula}
\usage{
latexImg(latex, markdown = T)
}
\description{
Creates a markdown link to the rendered form of a latex formula
}
