% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceWeb.R
\name{loadGithub}
\alias{loadGithub}
\title{Load an Rdata file from a URL}
\usage{
loadGithub(githubPath, branch = "master", envir = parent.frame(),
  token = NULL)
}
\arguments{
\item{githubPath}{character. username/repository/pathToFile}

\item{branch}{which branch to source from}

\item{envir}{the environment where the data should be loaded.}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
Load an Rdata file from a URL
}
\seealso{
\code{\link{loadURL}}, \code{\link{sourceGithub}}
}
