% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceWebAlt.R
\name{loadGithubAuth}
\alias{loadGithubAuth}
\title{Load RDA file from github.}
\usage{
loadGithubAuth(githubPath, branch = "master", envir = parent.frame(),
  token = NULL)
}
\description{
Loads data from github. Allows authentication. Max file size is 1 MB
}
