% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R
\name{loadURL}
\alias{loadURL}
\title{#' @export
"+" = function(x,y) {
    if(is.character(x) || is.character(y)) {
        return(paste(x , y, sep=""))
    } else {
        .Primitive("+")(x,y)
    }
}
Load an Rdata file from a URL}
\usage{
loadURL(url)
}
\arguments{
\item{url}{url of the Rdata file}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
#' @export
"+" = function(x,y) {
    if(is.character(x) || is.character(y)) {
        return(paste(x , y, sep=""))
    } else {
        .Primitive("+")(x,y)
    }
}
Load an Rdata file from a URL
}

