% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{prettifyLatexKable}
\alias{prettifyLatexKable}
\title{Prettify a latex table from knitr}
\usage{
prettifyLatexKable(table, columnLines = FALSE, headerLine = TRUE)
}
\description{
Makes latex tables coming out of knitr prettier. Requires booktabs latex package on the file
}
