% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{replace_assigns}
\alias{replace_assigns}
\title{Replace assignment operators}
\usage{
replace_assigns(file_in, file_out = file_in, good_assign = c("<-", "="))
}
\arguments{
\item{file_in}{input file}

\item{file_out}{output file}

\item{good_assign}{which operator should be used to replace}
}
\description{
Replace assignment operators
}
