% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R, R/ogbox2.R
\name{replaceElement}
\alias{replaceElement}
\title{Replaces elements of a vector based on the dictonary provided}
\usage{
replaceElement(vector, dictionary = NULL, NAreplace = NA)

replaceElement(vector, dictionary = NULL, NAreplace = NA)
}
\arguments{
\item{vector}{a vector}

\item{dictionary}{a named vector}

\item{NAreplace}{that to put instead of NAs}

\item{vector}{a vector}

\item{dictionary}{a named vector}

\item{NAreplace}{that to put instead of NAs}
}
\value{
A list with $newVector and $dictionary.

A list with $newVector and $dictionary.
}
\description{
Replaces elements of a vector based on the dictonary provided

Replaces elements of a vector based on the dictonary provided
}

