% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R
\name{roxygenTabular}
\alias{roxygenTabular}
\title{Roxygen table maker}
\usage{
roxygenTabular(df, col.names = TRUE, ...)
}
\arguments{
\item{df}{data.frame}

\item{col.names}{logica. If colnames should be included}

\item{...}{variables for format function}
}
\description{
Roxygen table maker
}
