% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{setDate}
\alias{setDate}
\title{Set version of a local package}
\usage{
setDate(date, pkg = ".")
}
\arguments{
\item{pkg}{Pakcage to set version of. Defaults to local directory}

\item{Date}{Character for the new date}
}
\description{
Set version of a local package
}
