% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{setVersion}
\alias{setVersion}
\title{Set version of a local package}
\usage{
setVersion(version, pkg = ".")
}
\arguments{
\item{version}{Character vector for the new version}

\item{pkg}{Pakcage to set version of. Defaults to local directory}
}
\description{
Set version of a local package
}
