% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{softDown}
\alias{softDown}
\title{Downloads soft file of a GSE}
\usage{
softDown(GSE, file, overwrite = FALSE)
}
\arguments{
\item{GSE}{GSE identifier of the dataset}

\item{file}{destination file}
}
\description{
Downloads a soft file.
}
