% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{softParser}
\alias{softParser}
\title{A parser for soft files}
\usage{
softParser(
  softFile,
  mergeFrame = c("intersect", "union"),
  n = NULL,
  expression = F
)
}
\arguments{
\item{softFile}{soft file to be parsed}

\item{mergeFrame}{if there are different data fields, what to do with them. not yet implemented}

\item{n}{how many samples are there in the GSE, if not provided, it will be acquired from GEO}

\item{expression}{should the expression data be returned}
}
\value{
A data.frame or a list depending on \code{expression}
}
\description{
Parses soft files in a flat file. Expression data inside the soft file can also be returned.
Returns a table of metadata if expression data is ommited. If not, returns a list. First element of the 
list is the metadata, the second element is a list of expression data that needs to be flattened independently
 since the expression tables can be different depending on the data
}
