% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceWebAlt.R
\name{sourceGithubAuth}
\alias{sourceGithubAuth}
\title{Source R files from github.}
\usage{
sourceGithubAuth(githubPath, lines = NULL, branch = "master",
  token = NULL)
}
\arguments{
\item{githubPath}{character. username/repository/pathToFile. Lines to excecute can be provided with github's #L1-L5 syntax. The function doesn't work with files that have ? # [ or ] in their addresses.}

\item{branch}{which branch to source from}

\item{line}{numbers to excecute. Providing lines in both here and with githubPath will result in an error.}
}
\description{
sources R files from github. Allows authentication. Max file size is 1 MB
}
\seealso{
\code{\link{loadURL}}, \code{\link{sourceGithub}}
}
