% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R
\name{trimHeadComment}
\alias{trimHeadComment}
\title{Trim first line comments}
\usage{
trimHeadComment(fileName, commentChar = "#", outFile = NULL)
}
\arguments{
\item{fileName}{name of the file}

\item{commentChar}{comment indicator}

\item{outFile}{File to save}
}
\value{
a textConnection object
}
\description{
Deletes the comments added to the first lines. Useful when comment char is 
only a comment char in the beginning of the file
}

