% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{whichGSE}
\alias{whichGSE}
\title{Gets GSE id of a GSM}
\usage{
whichGSE(GSM)
}
\arguments{
\item{gsm}{GSM identifier of the sample}
}
\value{
A character vectir of GSE identifiers
}
\description{
Gets GSE id of a GSM.
}
