% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceWeb.R
\name{loadURL}
\alias{loadURL}
\title{Load an Rdata file from a URL}
\usage{
loadURL(url, envir = parent.frame())
}
\arguments{
\item{url}{url of the Rdata file}

\item{envir}{the environment where the data should be loaded.}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
Load an Rdata file from a URL
}
