% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceWeb.R
\name{sourceGithub}
\alias{sourceGithub}
\title{Source R files from github}
\usage{
sourceGithub(githubPath, branch = "master")
}
\arguments{
\item{githubPath}{character. username/repository/pathToFile}

\item{branch}{which branch to source from}
}
\description{
sources R files from github
}
\seealso{
\code{\link{loadURL}}, \code{\link{sourceGithub}}
}

