% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogbox.R, R/ogbox2.R
\name{sourceGithub}
\alias{sourceGithub}
\title{Source R files from github}
\usage{
sourceGithub(user, repo, script)

sourceGithub(user, repo, script)
}
\arguments{
\item{user}{username}

\item{repo}{repository name}

\item{script}{name of the script}

\item{user}{username}

\item{repo}{repository name}

\item{script}{name of the script}
}
\description{
sources R files from github

sources R files from github
}
\details{
Does not require \code{user}, \code{repo} or \code{script} to be 
strings. Defaults to capitalized .R if extension not provided

Does not require \code{user}, \code{repo} or \code{script} to be 
strings. Defaults to capitalized .R if extension not provided
}

